formatting, naming, misc mainly for #16 fixed formatting (main problem was no end of file newline) moves classes to the file where the class content is changed event handler function calls to match convention
