+++ addons/core/CfgNotifications.hpp
