+++ addons/core/XEH_PREP.hpp
@@ -8,4 +8,5 @@ PREP(loadoutNotesGetLoadout);
 PREP(loadoutNotesName);
 PREP(loadoutNotesWepMags);
 PREP(orbat);
+PREP(teleport);
 PREP(teleportToSquad);
