+++ addons/core/config.cpp
@@ -16,3 +16,4 @@ class CfgPatches {
 
 #include "CfgEventHandlers.hpp"
 #include "CfgNotifications.hpp"
+#include "guiDefines\guiDefines.hpp"
