+++ addons/core/functions/fnc_teleport.sqf
@@ -1,0 +1,15 @@
+#include "script_component.hpp"
+
+/*
+ * Author: McKendrick
+ * TPD Teleport Script goes here.  Insert description here.
+ *
+ * Arguments:
+ * None
+ *
+ * Return Value:
+ * None
+ *
+ * Example:
+ * spawn sia3f_core_fnc_teleport
+*/
