+++ addons/main/script_macros.hpp
@@ -12,7 +12,7 @@
 #define GUI_THEME_ALPHA "(profileNamespace getVariable ['GUI_BCG_RGB_A',0.8])"
 #define GUI_THEME_COLOR {GUI_THEME_RGB_R,GUI_THEME_RGB_G,GUI_THEME_RGB_B,GUI_THEME_ALPHA}
 
-#define GET_CONFIG(var,default) (getMissionConfigValue [QEGVAR(configuration,var1), default])
+#define GET_CONFIG(var,default) (getMissionConfigValue [QEGVAR(configuration,var), default])
 
 // from ACE 3
 #ifdef DISABLE_COMPILE_CACHE
